+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 2.7.X - 2025-xx-xx
 - Feat:
 	- add `--primary`,`--secondary` and `--theme` css vars as default
+-Changes
+	- update header and footer default font size
 
 2.7.0 - 2025-10-23
 - Feat: