+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 	- add `utils.getStringAsElement` function
 	- add `utils.setValueFromKeys` function
 	- add `utils.flattenObj` function
+	- add `utils.resizeIframeOnContent` function
 -Changes
 	- update header and footer default font size
 	- update and improve `utils.request` function [breaking changes][WIP]